getinfo: Update to set paytxfee to 0.01
On mining pools with low minimum payouts, all payments below 0.01 failed.

Work-Around: If a min payout of 0.01 is set manually for the coin or the mining pool, payments works fine. This work-around is not needed anymore if this pull request is accepted.

I analysed and compared the code of TZC and SGL, I found this difference related to calculation of paytxfee. As TZC payments are working fine despite other wallet problems.

Remark: I don't think that this change will solve the problems described by ocminer in issue #1 but payments will be possible without work-around.